+++ CheckInSystem/Startup.cs
@@ -2,6 +2,7 @@
 using System.Diagnostics;
 using System.Windows;
 using CheckInSystem.CardReader;
+using CheckInSystem.Database;
 using CheckInSystem.Models;
 using CheckInSystem.ViewModels;
 using CheckInSystem.Views;
@@ -15,7 +16,7 @@ public class Startup
     {
         if (!EnsureDatabaseAvailable()) return false;
         ACR122U.StartReader();
-        ViewmodelBase.Employees = new ObservableCollection<Employee>(Employee.GetAllEmployees());
+        ViewmodelBase.Employees = new ObservableCollection<Employee>(DatabaseHelper.GetAllEmployees());
         ViewmodelBase.Groups =
             new ObservableCollection<Group>(Group.GetAllGroups(new List<Employee>(ViewmodelBase.Employees)));
         OpenEmployeeOverview();
